Commit for merge
This better be merged properly Boberto or I swear I'll rewatch KKHTA and find the perfect method to commit [extreme acts of uncivilized ultra-violent behavior] on you.

+++ src/main/java/com/hbm/blocks/machine/rbmk/RBMKDebrisRadiating.java
@@ -91,7 +91,7 @@ public class RBMKDebrisRadiating extends RBMKDebrisBurning {
 				res = 1;
 			
 			float eRads = rads;
-			eRads /= (float)res;
+			eRads /= res;
 			eRads /= (float)(len * len);
 			
 			ContaminationUtil.contaminate(e, HazardType.RADIATION, ContaminationType.CREATIVE, eRads);
